+++ src/plugins/devtools/devtools.module.ts
@@ -2,8 +2,11 @@ import { NgModule, ModuleWithProviders } from '@angular/core';
 import { NGXS_PLUGINS } from '../../symbols';
 import { NgxsDevtoolsOptions, NGXS_DEVTOOLS_OPTIONS } from './symbols';
 import { NgxsReduxDevtoolsPlugin } from './devtools.plugin';
+import { NgxsModule } from '../../module';
 
-@NgModule()
+@NgModule({
+  imports: [NgxsModule]
+})
 export class NgxsReduxDevtoolsPluginModule {
   static forRoot(options?: NgxsDevtoolsOptions): ModuleWithProviders {
     return {