+++ src/plugins/devtools/symbols.ts
@@ -6,7 +6,15 @@ import { InjectionToken } from '@angular/core';
 export interface NgxsDevtoolsExtension {
   init(state);
   send(action: string, state?: any);
-  subscribe(fn: (message: string) => void);
+  subscribe(fn: (message: NgxsDevtoolsAction) => void);
+}
+
+export interface NgxsDevtoolsAction {
+  type: string;
+  payload: any;
+  state: any;
+  id: number;
+  source: string;
 }
 
 export interface NgxsDevtoolsOptions {
@@ -14,6 +22,21 @@ export interface NgxsDevtoolsOptions {
    * Whether the dev tools is enabled or note. Useful for setting during production.
    */
   disabled: boolean;
+
+  /**
+   * Max number of entiries to keep.
+   */
+  maxAge?: number;
+
+  /**
+   * Reformat actions before sending to dev tools
+   */
+  actionSanitizer(action: any);
+
+  /**
+   * Reformat state before sending to devtools
+   */
+  stateSanitizer(state: any);
 }
 
 export const NGXS_DEVTOOLS_OPTIONS = new InjectionToken('NGXS_DEVTOOLS_OPTIONS');